fix payouts list
